Fix up pion-webrtc projects.
Remove fsnotify, which was a duplicate of gofsnotify. (These came as
dependencies of github.com/onsi/gomega, which for some reason has both
	github.com/fsnotify/fsnotify v1.4.7 // indirect
	gopkg.in/fsnotify.v1 v1.4.7 // indirect
in its go.mod file.)

Remove goxsync. It appeared as a dependency of gomega
	golang.org/x/sync v0.0.0-20180314180146-1d60e4601c6f // indirect
but everything compiles fine without it.

Add the missing goxtools.

Add go_lib_deps of projects that don't use go.mod.

Add go_lib_install of modules that are not at the top level of their
repo.

Upgrade goxnet to da137c7871d7 (which allows removing a formerly
nightly-only case, because the new commit is even newer than that one).

Upgrade goxsys to fae7ac547cb7.